% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogReturns.R
\name{logReturns}
\alias{logReturns}
\title{logReturns
Función que devuelve una matriz con los retornos logarítmicos de activos financieros, cuyos precios entran por parámetro.}
\usage{
logReturns(precios, tickers)
}
\arguments{
\item{precios}{Matriz o data frame de precios de activos. Cada activo debe estar representado por una columna.}

\item{tickers}{Vector de los tickers de los que se sacan las rentabilidades.}
}
\value{
Matriz con los retornos logaritmicos de los activos.
}
\description{
logReturns
Función que devuelve una matriz con los retornos logarítmicos de activos financieros, cuyos precios entran por parámetro.
}
\examples{
precios <- PrecioAcciones(...)
rentab <- logReturns(precios,c("HSY","AAPL","AMZN")) 
}
